scheduling: cron 表达式展开器只留一份实现
体检脚本问「这些槽发生了没有」，即将加的 dashboard 面板问「今天本该跑哪些」——
两边用同一个展开器，否则两个答案会各自漂移而看上去都对。
parse_cron_slots 从 ops/host/cron_health_check.py 搬进 clawock.scheduling，
ops 侧改成引用。
